Refactor model adapter #16
